+++ CHANGELOG.md
@@ -3,6 +3,11 @@
 # Change Log
 
 ## [Unreleased]
+
+### Fixed
+
+- Add cc-instanceinator compatibility
+
 ## [1.6.2] 2024-06-12
 ## [1.6.1] 2024-06-10